Remove generic constraints
This commit removes ID generic type constraints.
Unfortunately Go is a rather sad language which favours runtime
introspection instead of build time type safety.

The existing ID constraint was a hack which was put in place in hope of
making up for the lack of sum types. It did work pretty well, but made
code a bit verbose, which wouldnt be a huge deal, but the GO compiler is
very sad with type inference which makes things worse, still.

We are opting for runtime roulette with this commit and maybe revisiting
the type safety if Go introduces sum types.

Signed-off-by: Milos Gajdos <milosthegajdos@gmail.com>